refactor: move searchStreams.ts and extract new method
