+++ packages/client/src/StreamrClient.ts
@@ -22,7 +22,7 @@ import { StreamrClientEventEmitter, StreamrClientEvents } from './events'
 import { EthereumAddress, ProxyDirection, StreamID, StreamMessage } from 'streamr-client-protocol'
 import { MessageStream, MessageStreamOnMessage } from './subscribe/MessageStream'
 import { Stream, StreamProperties } from './Stream'
-import { SearchStreamsPermissionFilter } from './searchStreams'
+import { SearchStreamsPermissionFilter } from './registry/searchStreams'
 import { PermissionAssignment, PermissionQuery } from './permission'
 import { MetricsPublisher } from './MetricsPublisher'
 import { MessageMetadata } from './index-exports'
