+++ packages/client/test/unit/SearchStreams.test.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata'
 import { BigNumber } from '@ethersproject/bignumber'
 import { StreamID, StreamIDUtils, toStreamID } from 'streamr-client-protocol'
-import { SearchStreamsResultItem } from '../../src/searchStreams'
+import { SearchStreamsResultItem } from '../../src/registry/searchStreams'
 import { StreamRegistry } from '../../src/registry/StreamRegistry'
 import { collect } from '../../src/utils/GeneratorUtils'
 import { createMockAddress } from '../test-utils/utils'
